requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
openpyxl==3.1.5
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
Flask-WTF==1.2.1
gunicorn==22.0.0
